[WIP] Make a separate base quantum tape class (#829)
* separate base quantum tape from quantum tape

* pylint and black

* rename QuantumTape -> JacobianTape and BaseQuantumTape -> QuantumTape

* Update pennylane/tape/tapes/jacobian_tape.py


* implement changes from josh
