+++ tests/tape/interfaces/test_qnode_autograd.py
@@ -16,7 +16,7 @@ import pytest
 from pennylane import numpy as np
 
 import pennylane as qml
-from pennylane.tape import QuantumTape, qnode, QNode, QubitParamShiftTape
+from pennylane.tape import JacobianTape, qnode, QNode, QubitParamShiftTape
 
 
 @pytest.mark.parametrize(
@@ -121,7 +121,7 @@ class TestQNode:
 
     def test_jacobian(self, dev_name, diff_method, mocker, tol):
         """Test jacobian calculation"""
-        spy = mocker.spy(QuantumTape, "jacobian")
+        spy = mocker.spy(JacobianTape, "jacobian")
         a = np.array(0.1, requires_grad=True)
         b = np.array(0.2, requires_grad=True)
 
@@ -153,7 +153,7 @@ class TestQNode:
 
     def test_jacobian_no_evaluate(self, dev_name, diff_method, mocker, tol):
         """Test jacobian calculation when no prior circuit evaluation has been performed"""
-        spy = mocker.spy(QuantumTape, "jacobian")
+        spy = mocker.spy(JacobianTape, "jacobian")
         a = np.array(0.1, requires_grad=True)
         b = np.array(0.2, requires_grad=True)
 
@@ -189,7 +189,7 @@ class TestQNode:
         if diff_method == "backprop":
             pytest.skip("Test does not support backprop")
 
-        spy = mocker.spy(QuantumTape, "numeric_pd")
+        spy = mocker.spy(JacobianTape, "numeric_pd")
 
         a = np.array([0.1, 0.2], requires_grad=True)
 
@@ -255,7 +255,7 @@ class TestQNode:
         expected = [-np.sin(a) + np.sin(a) * np.sin(b)]
         assert np.allclose(res, expected, atol=tol, rtol=0)
 
-        # QuantumTape.numeric_pd has been called only once
+        # JacobianTape.numeric_pd has been called only once
         assert len(spy.call_args_list) == 1
 
         # trainability also updates on evaluation
@@ -352,7 +352,7 @@ class TestQNode:
                 theta, phi, lam = self.data
                 wires = self.wires
 
-                with QuantumTape() as tape:
+                with JacobianTape() as tape:
                     qml.Rot(lam, theta, -lam, wires=wires)
                     qml.PhaseShift(phi + lam, wires=wires)
 
@@ -373,8 +373,8 @@ class TestQNode:
         if diff_method == "finite-diff":
             assert circuit.qtape.trainable_params == {1, 2, 3, 4}
         elif diff_method == "backprop":
-            # For a backprop device, no interface wrapping is performed, and QuantumTape.jacobian()
-            # is never called. As a result, QuantumTape.trainable_params is never set --- the ML
+            # For a backprop device, no interface wrapping is performed, and JacobianTape.jacobian()
+            # is never called. As a result, JacobianTape.trainable_params is never set --- the ML
             # framework uses its own backprop logic and its own bookkeeping re: trainable parameters.
             assert circuit.qtape.trainable_params == {0, 1, 2, 3, 4}
